MOD: Split DatabaseAddRemoveController.java into DatabaseAddController.java and DatabaseRemoveController.java
