+++ src/controllers/StudentController.java
@@ -4,14 +4,14 @@
 
 package controllers;
 
-import database.DatabaseAddRemoveController;
+import database.DatabaseAddController;
 import models.Student;
 
 public class StudentController extends UserController {
 //	Methods:
 	public String createStudentCardSaveAndReturnId(String[] formArray) {
 		var student = new Student(formArray[0], formArray[1], formArray[2]);
-		new DatabaseAddRemoveController().addStudentToList(student);
+		new DatabaseAddController().addStudentToList(student);
 		return student.getUserId();
 	}
 }
